+++ integrations/appboy/lib/index.js
@@ -43,6 +43,14 @@ var Appboy = (module.exports = integration('Appboy')
   .tag(
     'v2',
     '<script src="https://js.appboycdn.com/web-sdk/2.4/appboy.min.js">'
+  )
+  .tag(
+    'v2.7',
+    '<script src="https://js.appboycdn.com/web-sdk/2.7/appboy.min.js">'
+  )
+  .tag(
+    'v3.1',
+    '<script src="https://js.appboycdn.com/web-sdk/3.1/appboy.min.js">'
   ));
 
 Appboy.prototype.appboyInitialize = function(userId, options, config) {
@@ -56,7 +64,9 @@ Appboy.prototype.appboyInitialize = function(userId, options, config) {
 };
 
 Appboy.prototype.initialize = function() {
-  if (appboyUtil.isVersionTwo(this.options)) {
+  if (appboyUtil.isMajorVersionThree(this.options)) {
+    this.initializeV3();
+  } else if (appboyUtil.isMajorVersionTwo(this.options)) {
     this.initializeV2();
   } else {
     this.initializeV1();
@@ -160,7 +170,33 @@ Appboy.prototype.initializeV2 = function() {
     this.appboyInitialize(userId, options, config);
   }
 
-  this.load('v2', this.ready);
+  var versionTag = Number(options.version) === 2.7 ? 'v2.7' : 'v2';
+  this.load(versionTag, this.ready);
+};
+
+Appboy.prototype.initializeV3 = function() {
+  var options = this.options;
+  var userId = this.analytics.user().id();
+
+  /* eslint-disable */
+  +function(a,p,P,b,y){
+    a.appboy={};
+    a.appboyQueue=[];
+    for(var s="DeviceProperties Card Card.prototype.dismissCard Card.prototype.removeAllSubscriptions Card.prototype.removeSubscription Card.prototype.subscribeToClickedEvent Card.prototype.subscribeToDismissedEvent Banner CaptionedImage ClassicCard ControlCard ContentCards ContentCards.prototype.getUnviewedCardCount Feed Feed.prototype.getUnreadCardCount ControlMessage InAppMessage InAppMessage.SlideFrom InAppMessage.ClickAction InAppMessage.DismissType InAppMessage.OpenTarget InAppMessage.ImageStyle InAppMessage.Orientation InAppMessage.TextAlignment InAppMessage.CropType InAppMessage.prototype.closeMessage InAppMessage.prototype.removeAllSubscriptions InAppMessage.prototype.removeSubscription InAppMessage.prototype.subscribeToClickedEvent InAppMessage.prototype.subscribeToDismissedEvent FullScreenMessage ModalMessage HtmlMessage SlideUpMessage User User.Genders User.NotificationSubscriptionTypes User.prototype.addAlias User.prototype.addToCustomAttributeArray User.prototype.getUserId User.prototype.incrementCustomUserAttribute User.prototype.removeFromCustomAttributeArray User.prototype.setAvatarImageUrl User.prototype.setCountry User.prototype.setCustomLocationAttribute User.prototype.setCustomUserAttribute User.prototype.setDateOfBirth User.prototype.setEmail User.prototype.setEmailNotificationSubscriptionType User.prototype.setFirstName User.prototype.setGender User.prototype.setHomeCity User.prototype.setLanguage User.prototype.setLastKnownLocation User.prototype.setLastName User.prototype.setPhoneNumber User.prototype.setPushNotificationSubscriptionType InAppMessageButton InAppMessageButton.prototype.removeAllSubscriptions InAppMessageButton.prototype.removeSubscription InAppMessageButton.prototype.subscribeToClickedEvent display display.automaticallyShowNewInAppMessages display.destroyFeed display.hideContentCards display.showContentCards display.showFeed display.showInAppMessage display.toggleContentCards display.toggleFeed changeUser destroy getDeviceId initialize isPushBlocked isPushGranted isPushPermissionGranted isPushSupported logCardClick logCardDismissal logCardImpressions logContentCardsDisplayed logCustomEvent logFeedDisplayed logInAppMessageButtonClick logInAppMessageClick logInAppMessageHtmlClick logInAppMessageImpression logPurchase openSession registerAppboyPushMessages removeAllSubscriptions removeSubscription requestContentCardsRefresh requestFeedRefresh requestImmediateDataFlush resumeWebTracking setLogger stopWebTracking subscribeToContentCardsUpdates subscribeToFeedUpdates subscribeToInAppMessage subscribeToNewInAppMessages toggleAppboyLogging trackLocation unregisterAppboyPushMessages wipeData".split(" "),i=0;i<s.length;i++){for(var m=s[i],k=a.appboy,l=m.split("."),j=0;j<l.length-1;j++)k=k[l[j]];k[l[j]]=(new Function("return function "+m.replace(/\./g,"_")+"(){window.appboyQueue.push(arguments); return true}"))()}
+    window.appboy.getCachedContentCards=function(){return new window.appboy.ContentCards};
+    window.appboy.getCachedFeed=function(){return new window.appboy.Feed};
+    window.appboy.getUser=function(){return new window.appboy.User};
+  }(window,document,'script');
+  /* eslint-enable */
+
+  if (appboyUtil.shouldOpenSession(userId, options)) {
+    this.hasBeenInitialized = true;
+    var config = appboyUtil.getConfig(options);
+    this.initializeTester(options.apiKey, config);
+    this.appboyInitialize(userId, options, config);
+  }
+
+  this.load('v3.1', this.ready);
 };
 
 /**