+++ integrations/gtag/HISTORY.md
@@ -1,3 +1,9 @@
+1.0.1 / 2020-11-02
+==================
+
+  * Add `window.gtag('js', new Date())` command to initialize gtag event tracking.
+
+
 1.0.0 / 2020-01-22
 ==================